refactor: old recurring payments correct use of providers
