+++ lib/providers/transactions_provider.dart
@@ -3,7 +3,6 @@ import 'package:riverpod_annotation/riverpod_annotation.dart';
 import '../model/bank_account.dart';
 import '../model/recurring_transaction.dart';
 import '../model/transaction.dart';
-import '../services/database/repositories/recurring_transactions_repository.dart';
 import '../services/database/repositories/transactions_repository.dart';
 import 'accounts_provider.dart';
 import 'budgets_provider.dart';
@@ -84,15 +83,6 @@ class EndDate extends _$EndDate {
   DateTime? build() => null;
 
   void setDate(DateTime? date) => state = date;
-}
-
-@Riverpod(keepAlive: true)
-class SelectedRecurringTransactionUpdate
-    extends _$SelectedRecurringTransactionUpdate {
-  @override
-  RecurringTransaction? build() => null;
-
-  void setValue(RecurringTransaction? value) => state = value;
 }
 
 // Amount total for the transactions filtered
@@ -342,134 +332,6 @@ Future<List<Transaction>> monthlyTransactions(Ref ref) async {
       .selectAll(dateRangeStart: firstDayOfMonth, dateRangeEnd: now);
 
   return transactions;
-}
-
-@Riverpod(keepAlive: true)
-class RecurringTransactionsNotifier extends _$RecurringTransactionsNotifier {
-  @override
-  Future<List<RecurringTransaction>> build() {
-    return _getRecurringTransactions();
-  }
-
-  Future<List<RecurringTransaction>> _getRecurringTransactions() async {
-    final transactions = await ref
-        .read(recurringTransactionRepositoryProvider)
-        .selectAllActive();
-    return transactions;
-  }
-
-  Future<RecurringTransaction?> create(
-    num amount,
-    String label,
-    TransactionType type,
-  ) async {
-    state = const AsyncLoading();
-
-    final date = ref.read(selectedDateProvider);
-    final toDate = ref.read(endDateProvider);
-    final bankAccount = ref.read(selectedBankAccountProvider)!;
-    final category = ref.read(selectedCategoryProvider);
-    final recurrency = ref.read(intervalProvider);
-
-    RecurringTransaction transaction = RecurringTransaction(
-      amount: amount,
-      fromDate: date,
-      toDate: toDate,
-      note: label,
-      type: type,
-      idBankAccount: bankAccount.id!,
-      idCategory: category!.id!,
-      recurrency: recurrency,
-      createdAt: date,
-      updatedAt: date,
-      lastInsertion: date,
-    );
-
-    // Here we need the recurringTransaction just inserted, to get and return a model with also his ID
-    RecurringTransaction? insertedTransaction;
-
-    state = await AsyncValue.guard(() async {
-      insertedTransaction = await ref
-          .read(recurringTransactionRepositoryProvider)
-          .insert(transaction);
-
-      // check if fromDate is today, and add the first recurrence of the transaction
-      DateTime now = DateTime.now();
-
-      if (date.year == now.year &&
-          date.month == now.month &&
-          date.day == now.day) {
-        final transaction = Transaction(
-          date: date,
-          amount: amount,
-          type: type,
-          note: label,
-          idBankAccount: bankAccount.id!,
-          idCategory: category.id!,
-          idRecurringTransaction: insertedTransaction!.id,
-          recurring: true,
-        );
-        await ref.read(transactionsRepositoryProvider).insert(transaction);
-      }
-      return await _getRecurringTransactions();
-    });
-
-    return insertedTransaction;
-  }
-
-  Future<void> updateTransaction(num amount, String label) async {
-    final bankAccount = ref.read(selectedBankAccountProvider)!;
-    final recurrency = ref.read(intervalProvider);
-    final category = ref.read(selectedCategoryProvider);
-
-    final RecurringTransaction transaction = ref
-        .read(selectedRecurringTransactionUpdateProvider)!
-        .copy(
-          fromDate: ref.read(selectedDateProvider),
-          toDate: ref.read(endDateProvider),
-          recurrency: recurrency,
-          amount: amount,
-          note: label,
-          idBankAccount: bankAccount.id!,
-          idCategory: category?.id,
-          updatedAt: DateTime.now(),
-        );
-
-    state = const AsyncLoading();
-    state = await AsyncValue.guard(() async {
-      await ref
-          .read(recurringTransactionRepositoryProvider)
-          .updateItem(transaction);
-      return await _getRecurringTransactions();
-    });
-  }
-
-  Future<void> transactionSelect(RecurringTransaction transaction) async {
-    ref
-        .read(selectedRecurringTransactionUpdateProvider.notifier)
-        .setValue(transaction);
-    ref.read(selectedRecurringPayProvider.notifier).setValue(true);
-    ref.read(selectedCategoryProvider.notifier).state = ref
-        .read(categoriesProvider)
-        .value!
-        .firstWhere((element) => element.id == transaction.idCategory);
-    ref.read(selectedBankAccountProvider.notifier).state = ref
-        .read(accountsProvider)
-        .value!
-        .firstWhere((element) => element.id == transaction.idBankAccount);
-    ref.read(intervalProvider.notifier).setValue(transaction.recurrency);
-    ref.read(endDateProvider.notifier).setDate(transaction.toDate);
-  }
-
-  Future<void> delete(int transactionId) async {
-    state = const AsyncLoading();
-    state = await AsyncValue.guard(() async {
-      await ref
-          .read(recurringTransactionRepositoryProvider)
-          .deleteById(transactionId);
-      return _getRecurringTransactions();
-    });
-  }
 }
 
 @riverpod
